fix: develop을 merge하면서 논리적으로 필요한 변경사항 반영

+++ src/main/java/com/samhap/kokomen/interview/service/dto/InterviewResultResponse.java
@@ -8,7 +8,7 @@ public record InterviewResultResponse(
         List<FeedbackResponse> feedbacks,
         String totalFeedback,
         Integer totalScore,
-        Integer interviewLikeCount,
+        Long interviewLikeCount,
         Boolean interviewAlreadyLiked,
         Integer userCurScore,
         Integer userPrevScore
